Split ConversationMessageBubble in two files to reduce functions count
