+++ src/com/android/messaging/ui/conversation/messages/ui/message/ConversationMessageBubble.kt
@@ -1,23 +1,15 @@
 package com.android.messaging.ui.conversation.messages.ui.message
 
-import androidx.compose.animation.animateColorAsState
-import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.widthIn
-import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Surface
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.CompositionLocalProvider
-import androidx.compose.runtime.getValue
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.clip
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.text.style.TextOverflow
 import androidx.compose.ui.tooling.preview.PreviewLightDark
@@ -40,11 +32,8 @@ import com.android.messaging.ui.conversation.preview.previewVCardPart
 import com.android.messaging.ui.conversation.preview.previewVideoPart
 import kotlinx.collections.immutable.persistentListOf
 
-private val MESSAGE_BUBBLE_MEDIA_SECTION_SPACING = 8.dp
-private val MESSAGE_BUBBLE_MEDIA_TEXT_PADDING = 12.dp
 private val MESSAGE_BUBBLE_TEXT_HORIZONTAL_PADDING = 16.dp
 private val MESSAGE_BUBBLE_TEXT_VERTICAL_PADDING = 12.dp
-private const val MESSAGE_SELECTION_MEDIA_OVERLAY_ALPHA = 0.2f
 
 @Composable
 internal fun ConversationMessageBubble(
@@ -107,65 +96,6 @@ internal fun ConversationMessageBubble(
 }
 
 @Composable
-private fun ConversationMessageAttachmentOnlyBubble(
-    modifier: Modifier,
-    layout: ConversationMessageLayout,
-    message: ConversationMessageUiModel,
-    isSelected: Boolean,
-    isSelectionMode: Boolean,
-    onAttachmentClick: OnConversationAttachmentClick,
-    onExternalUriClick: (String) -> Unit,
-    onMessageLongClick: () -> Unit,
-) {
-    ConversationMessageAttachmentOnlyContainer(
-        modifier = modifier,
-        bubbleShape = layout.bubbleShape,
-        message = message,
-        isSelected = isSelected,
-    ) {
-        ConversationMessageAttachmentBubbleContent(
-            modifier = Modifier.fillMaxWidth(),
-            layout = layout,
-            message = message,
-            isSelected = isSelected,
-            isSelectionMode = isSelectionMode,
-            onAttachmentClick = onAttachmentClick,
-            onExternalUriClick = onExternalUriClick,
-            onMessageLongClick = onMessageLongClick,
-        )
-    }
-}
-
-@Composable
-private fun ConversationMessageAttachmentSurfaceBubble(
-    modifier: Modifier,
-    layout: ConversationMessageLayout,
-    isSelected: Boolean,
-    message: ConversationMessageUiModel,
-    isSelectionMode: Boolean,
-    onAttachmentClick: OnConversationAttachmentClick,
-    onExternalUriClick: (String) -> Unit,
-    onMessageLongClick: () -> Unit,
-) {
-    ConversationMessageBubbleSurface(
-        modifier = modifier,
-        isSelected = isSelected,
-        message = message,
-        layout = layout,
-    ) {
-        ConversationMessageAttachmentBubbleContent(
-            layout = layout,
-            message = message,
-            isSelected = isSelected,
-            isSelectionMode = isSelectionMode,
-            onAttachmentClick = onAttachmentClick,
-            onExternalUriClick = onExternalUriClick,
-            onMessageLongClick = onMessageLongClick,
-        )
-    }
-}
-
-@Composable
 private fun ConversationMessageTextSurfaceBubble(
     modifier: Modifier,
     layout: ConversationMessageLayout,
@@ -197,7 +127,7 @@ private fun ConversationMessageTextSurfaceBubble(
 }
 
 @Composable
-private fun ConversationMessageBubbleSurface(
+internal fun ConversationMessageBubbleSurface(
     modifier: Modifier = Modifier,
     isSelected: Boolean,
     message: ConversationMessageUiModel,
@@ -222,44 +152,6 @@ private fun ConversationMessageBubbleSurface(
             LocalConversationMessageLinkColor provides contentColor,
         ) {
             bubbleContent()
-        }
-    }
-}
-
-@Composable
-private fun ConversationMessageAttachmentOnlyContainer(
-    modifier: Modifier = Modifier,
-    bubbleShape: RoundedCornerShape,
-    message: ConversationMessageUiModel,
-    isSelected: Boolean,
-    content: @Composable () -> Unit,
-) {
-    val overlayColor by animateColorAsState(
-        targetValue = when {
-            isSelected -> {
-                messageBubbleColor(
-                    message = message,
-                    isSelected = true,
-                ).copy(alpha = MESSAGE_SELECTION_MEDIA_OVERLAY_ALPHA)
-            }
-
-            else -> Color.Transparent
-        },
-        label = "conversationMessageSelectionOverlayColor",
-    )
-
-    Box(
-        modifier = modifier.clip(shape = bubbleShape),
-    ) {
-        content()
-
-        if (overlayColor != Color.Transparent) {
-            Box(
-                modifier = Modifier
-                    .fillMaxSize()
-                    .clip(shape = bubbleShape)
-                    .background(color = overlayColor),
-            )
         }
     }
 }
@@ -322,94 +214,6 @@ private fun ConversationMessageTextBubbleContent(
 }
 
 @Composable
-private fun ConversationMessageAttachmentBubbleContent(
-    modifier: Modifier = Modifier,
-    layout: ConversationMessageLayout,
-    message: ConversationMessageUiModel,
-    isSelected: Boolean,
-    isSelectionMode: Boolean,
-    onAttachmentClick: OnConversationAttachmentClick,
-    onExternalUriClick: (String) -> Unit,
-    onMessageLongClick: () -> Unit,
-) {
-    val content = layout.content
-    val hasHeader = layout.showSender || !content.subjectText.isNullOrBlank()
-    val hasBodyText = !content.bodyText.isNullOrBlank()
-
-    Column(
-        modifier = modifier.fillMaxWidth(),
-    ) {
-        ConversationMessageSender(
-            modifier = Modifier.padding(
-                start = MESSAGE_BUBBLE_MEDIA_TEXT_PADDING,
-                top = MESSAGE_BUBBLE_MEDIA_TEXT_PADDING,
-                end = MESSAGE_BUBBLE_MEDIA_TEXT_PADDING,
-                bottom = conversationMessageSenderBottomPadding(content),
-            ),
-            color = messageSenderColor(
-                message = message,
-                isSelected = isSelected,
-            ),
-            senderDisplayName = message.senderDisplayName,
-            showSender = layout.showSender,
-        )
-
-        ConversationMessageSubject(
-            modifier = Modifier.padding(
-                start = MESSAGE_BUBBLE_MEDIA_TEXT_PADDING,
-                top = conversationMessageSubjectTopPadding(showSender = layout.showSender),
-                end = MESSAGE_BUBBLE_MEDIA_TEXT_PADDING,
-                bottom = MESSAGE_BUBBLE_MEDIA_SECTION_SPACING,
-            ),
-            subjectText = content.subjectText,
-        )
-
-        ConversationMessageAttachments(
-            attachmentSections = content.attachmentSections,
-            hasTextAboveVisualAttachments = hasHeader,
-            hasTextBelowVisualAttachments = hasBodyText,
-            isIncoming = message.isIncoming,
-            isSelectionMode = isSelectionMode,
-            useStandaloneAudioAttachmentBg = false,
-            onAttachmentClick = onAttachmentClick,
-            onExternalUriClick = onExternalUriClick,
-            onMessageLongClick = onMessageLongClick,
-        )
-
-        content.bodyText?.let { bodyText ->
-            ConversationMessageText(
-                modifier = Modifier.padding(
-                    start = MESSAGE_BUBBLE_MEDIA_TEXT_PADDING,
-                    top = MESSAGE_BUBBLE_MEDIA_SECTION_SPACING,
-                    end = MESSAGE_BUBBLE_MEDIA_TEXT_PADDING,
-                    bottom = MESSAGE_BUBBLE_MEDIA_TEXT_PADDING,
-                ),
-                text = bodyText,
-                style = MaterialTheme.typography.bodyLarge,
-                onExternalUriClick = onExternalUriClick,
-                onMessageLongClick = onMessageLongClick,
-            )
-        }
-    }
-}
-
-private fun conversationMessageSenderBottomPadding(
-    content: ConversationMessageContent,
-): Dp {
-    return when {
-        content.subjectText.isNullOrBlank() -> 6.dp
-        else -> MESSAGE_BUBBLE_MEDIA_SECTION_SPACING
-    }
-}
-
-private fun conversationMessageSubjectTopPadding(showSender: Boolean): Dp {
-    return when {
-        showSender -> 0.dp
-        else -> MESSAGE_BUBBLE_MEDIA_TEXT_PADDING
-    }
-}
-
-@Composable
 private fun ConversationMessageBody(
     content: ConversationMessageContent,
     isIncoming: Boolean,
@@ -443,7 +247,7 @@ private fun ConversationMessageBody(
 }
 
 @Composable
-private fun ConversationMessageSubject(
+internal fun ConversationMessageSubject(
     subjectText: String?,
     modifier: Modifier = Modifier,
 ) {
@@ -459,7 +263,7 @@ private fun ConversationMessageSubject(
 }
 
 @Composable
-private fun ConversationMessageSender(
+internal fun ConversationMessageSender(
     modifier: Modifier = Modifier,
     color: Color,
     senderDisplayName: String?,
@@ -480,7 +284,7 @@ private fun ConversationMessageSender(
 }
 
 @Composable
-private fun messageBubbleColor(
+internal fun messageBubbleColor(
     message: ConversationMessageUiModel,
     isSelected: Boolean,
 ): Color {
@@ -504,7 +308,7 @@ private fun messageBubbleContentColor(
 }
 
 @Composable
-private fun messageSenderColor(
+internal fun messageSenderColor(
     message: ConversationMessageUiModel,
     isSelected: Boolean,
 ): Color {
